+++ sbin/generateViewerFileForStaticData.sh
@@ -22,7 +22,7 @@ source ${SCRIPTPATH}/configure-environment.bash
 VIEWERSRC=${SCRIPTPATH}/../dist/static/myshpviewer/de/index.viewer.full.html
 
 if [ ! -d "${EXPORTDIR}" ]; then
-    dofail "USAGE: prepareViewerFileForStaticData.sh EXPORTDIR\nFATAL: $EXPORTDIR must exists" 1
+    dofail "USAGE: generateViewerFileForStaticData.sh EXPORTDIR\nFATAL: $EXPORTDIR must exists" 1
     exit 1
 fi