+++ src/frontend/app/app.common.module.ts
@@ -9,8 +9,8 @@ import {SearchFormUtils} from '@dps/mycms-frontend-commons/dist/angular-commons/
 import {GenericAppService} from '@dps/mycms-commons/dist/commons/services/generic-app.service';
 import {SearchParameterUtils} from '@dps/mycms-commons/dist/search-commons/services/searchparameter.utils';
 import {SectionsModule} from './sections/sections.module';
-import {PDocDataService} from '@dps/mycms-commons/dist/pdoc-commons/services/pdoc-data.service';
-import {PDocDataStore} from '@dps/mycms-commons/dist/pdoc-commons/services/pdoc-data.store';
+import {StaticPagesDataService} from '@dps/mycms-commons/dist/pdoc-commons/services/staticpages-data.service';
+import {StaticPagesDataStore} from '@dps/mycms-commons/dist/pdoc-commons/services/staticpages-data.store';
 import {BrowserModule} from '@angular/platform-browser';
 import {ErrorPageComponent} from './components/errorpage/errorpage.component';
 import {AngularCommonsModule} from '@dps/mycms-frontend-commons/dist/angular-commons/angular-commons.module';
@@ -34,6 +34,8 @@ import {FallbackHttpClient} from './services/fallback-http-client';
 import {NgModule} from '@angular/core';
 import {AppCommonRoutingModule} from './app.common.router';
 import {CommonDocRoutingService} from '@dps/mycms-frontend-commons/dist/frontend-cdoc-commons/services/cdoc-routing.service';
+import {PDocDataService} from '@dps/mycms-commons/dist/pdoc-commons/services/pdoc-data.service';
+import {PDocDataStore} from '@dps/mycms-commons/dist/pdoc-commons/services/pdoc-data.store';
 
 registerLocaleData(localeDe);
 
@@ -84,6 +86,8 @@ export function getAngulartics2Providers(): any {
         DynamicComponentService,
         PDocDataStore,
         PDocDataService,
+        StaticPagesDataStore,
+        StaticPagesDataService,
         SearchFormUtils,
         {provide: GenericTrackingService, useClass: TrackingService},
         AngularHtmlService,