+++ docs/CHANGELOG.md
@@ -2,6 +2,30 @@
  
 # Versions
 
+## 5.0.0
+- use commons 6.0.0
+- bumped up deps
+- added optional password-encryption of static-viewer-version
+- backend/frontend: add page-management
+
+### new features
+- added optional password-encryption of static-viewer-version
+- backend/frontend: add page-management
+
+### improvements
+- use commons 6.0.0
+- bumped up deps
+
+### bug fixes
+- none
+
+### breaking changes
+- use commons 6.0.0
+- bumped up deps
+- use new config
+- use new pdoc-config -> see [Migration-instructions](MIGRATION.md)
+
+
 ## 4.4.0 (2023-03-18)
 - bumped up to commons 5.7.0